Short and long data transfer test.
These appear to work reasonably well with an
occasional transient failure.
